Creating / updating categories

+++ view/admin/categories/list.php
@@ -7,7 +7,7 @@
 <?php
 function display_categories($categories) {
 	foreach ($categories as $category) {
-		echo '<li>' . $category['name'] . ' - <span class="btn-group"><a href="categories/edit/' . $category['id'] . '" class="btn btn-xs btn-info">Edit</a><a href="categories/delete/' . $category['id'] . '" class="btn btn-xs btn-danger">Delete</a></span>';
+		echo '<li>' . $category['name'] . ' - <span class="btn-group"><a href="categories/edit/' . $category['id'] . '" class="btn btn-xs btn-info">Edit</a><a href="categories/delete/' . $category['id'] . '" class="btn btn-xs btn-danger delete-confirm">Delete</a></span>';
 		if (count($category['children'])) {
 			echo '<ul>';
 			display_categories($category['children']);
